Test fixture DRY, build count for users
