+++ mig3/accounts/tests/test_useraccount.py
@@ -1,6 +1,8 @@
 from django.contrib.auth import get_user_model
 from django.contrib.auth.backends import ModelBackend
 
+from builds import models as builds
+
 UserAccount = get_user_model()
 
 DEFAULT_PARAMETERS = {"email": "user@example.com", "name": "Test User", "password": "TestPassword", "is_active": True}
@@ -52,3 +54,16 @@ def test_create_inactive_user(db):
     user = UserAccount.objects.create_user(**parameters)
     password = parameters.pop("password")
     assert not ModelBackend().authenticate(request=None, username=user.email, password=password)
+
+
+def test_build_count(build, another_version, user_account):
+    """Should have accurate build count."""
+    assert user_account.build_count == builds.Build.objects.filter(version__author=user_account).count()
+
+    # Reset ID for new record, then assign new number, and version to remove uniqueness constraints
+    build.id = None
+    build.number = "2"
+    build.version = another_version
+    build.save()
+
+    assert user_account.build_count == builds.Build.objects.filter(version__author=user_account).count()
